+++ src-tauri/Cargo.toml
@@ -22,4 +22,6 @@ tauri = { version = "2", features = [] }
 tauri-plugin-shell = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+surrealdb = "2.0.4"
+tokio = { version = "1.41.0", features = ["macros", "rt-multi-thread"] }
