+++ src/blocks/todo/todo-m.ts
@@ -1,0 +1,6 @@
+export type Task = {
+  task: string;
+  order: number;
+  name: string;
+  checked: boolean;
+};
